+++ packages/core/navigation-next/examples/13-async-load-layout-manager-with-view-controller.js
@@ -11,7 +11,8 @@ import {
   AsyncLayoutManagerWithViewController,
   NavigationProvider,
   SkeletonContainerView,
-  ViewRenderer,
+  ItemsRenderer,
+  SkeletonItem,
 } from '../src';
 
 import ContainerViews from './shared/views/container';
@@ -38,10 +39,22 @@ const AsyncProjectSwitch = asyncComponent({
       return new Promise(resolve => {
         setTimeout(() => {
           return resolve(ProjectSwitcher);
-        }, 10000);
+        }, 2000);
       });
     }),
   LoadingComponent: () => <ProjectSwitchSkeleton />,
+});
+
+const AsyncLinkItem = asyncComponent({
+  resolve: () =>
+    import('./shared/components').then(({ LinkItem }) => {
+      return new Promise(resolve => {
+        setTimeout(() => {
+          return resolve(LinkItem);
+        }, 2000);
+      });
+    }),
+  LoadingComponent: () => <SkeletonItem hasBefore />,
 });
 
 const AsyncDefaultGlobalNavigation = asyncComponent({
@@ -113,16 +126,16 @@ export default class App extends Component<
 
     return (
       <HashRouter>
-        <NavigationProvider
-          initialPeekViewId="root/index"
-          isDebugEnabled={isDebugEnabled}
-        >
+        <NavigationProvider isDebugEnabled={isDebugEnabled}>
           <AsyncLayoutManagerWithViewController
-            customComponents={{ ProjectSwitcher: AsyncProjectSwitch }}
+            customComponents={{
+              ProjectSwitcher: AsyncProjectSwitch,
+              LinkItem: AsyncLinkItem,
+            }}
             experimental_flyoutOnHover={isFlyoutAvailable}
             globalNavigation={AsyncDefaultGlobalNavigation}
             containerSkeleton={() => <SkeletonContainerView type={'product'} />}
-            viewRenderer={ViewRenderer}
+            itemsRenderer={ItemsRenderer}
             firstSkeletonToRender={'product'}
           >
             <div style={{ padding: 40 }}>