+++ packages/editor/adf-utils/src/validator/specs/tableCell_content.ts
@@ -2,15 +2,15 @@ export default {
   type: 'array',
   items: [
     [
-      'paragraph',
+      'paragraph_with_no_marks',
       'bulletList',
       'mediaSingle',
       'orderedList',
-      'heading',
+      'heading_with_no_marks',
       'panel',
       'blockquote',
       'rule',
-      'codeBlock_no_marks',
+      'codeBlock_with_no_marks',
       'mediaGroup',
       'applicationCard',
       'decisionList',