+++ packages/editor/editor-core/src/plugins/type-ahead/commands/dismiss.ts
@@ -1,6 +1,7 @@
 import { Command } from '../../../types';
 import { findTypeAheadQuery } from '../utils/find-query-mark';
 import { analyticsService } from '../../../analytics';
+import { pluginKey } from '../pm-plugins/main';
 
 export const dismissCommand = (): Command => (state, dispatch) => {
   const queryMark = findTypeAheadQuery(state);
@@ -17,6 +18,11 @@ export const dismissCommand = (): Command => (state, dispatch) => {
   }
 
   analyticsService.trackEvent('atlassian.editor.typeahead.dismiss');
+
+  const { typeAheadHandler } = pluginKey.getState(state);
+  if (typeAheadHandler && typeAheadHandler.dismiss) {
+    typeAheadHandler.dismiss(state);
+  }
 
   dispatch(
     state.tr.removeMark(start, end, markType).removeStoredMark(markType),