+++ packages/editor/editor-wikimarkup-transformer/src/__tests__/unit/parser/text-effects.ts
@@ -56,6 +56,10 @@ https://app.datadoghq.com/screen/282018/product-fabric-adf-service?tv_mode=true#
       '[CS-576] should allow {color} macro in formatters',
       'This is _{color:red}*Strong Red and Italic*{color}_',
     ],
+    [
+      'should not apply strong when ending line is two strong symbols',
+      '*not valid strong **',
+    ],
   ];
 
   for (const [testCaseDescription, markup] of testCases) {