+++ packages/elements/user-picker/src/components/utils.ts
@@ -49,5 +49,11 @@ export const usersToOptions = memoizeOne((defaultValue: UserValue) => {
   return userToOption(defaultValue);
 });
 
-export const getAvatarSize = (appearance: string): 'small' | 'xsmall' =>
-  appearance === 'normal' ? 'small' : 'xsmall';
+export const getAvatarSize = (
+  appearance: string,
+): 'small' | 'xsmall' | 'medium' =>
+  appearance === 'normal'
+    ? 'small'
+    : appearance === 'big'
+    ? 'medium'
+    : 'xsmall';